+++ src/core/shared/interfaces/Marker.ts
@@ -7,4 +7,5 @@ export interface Marker {
     scale?: Vector3;
     maxDistance?: number;
     uid?: string;
+    dimension?: number;
 }